backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
torchaudio==0.16.0
diffusers==0.21.4
transformers==4.34.0
peft==0.7.0
safetensors==0.4.0
pillow==10.1.0
pydantic==2.5.0
python-multipart==0.0.6
